+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.4.0
+Add option to have a note created for each folder (for folder note plugins).
+
+Ensure that full hierarchy of folders is created, not just the first level.
+
+Add title and aliases for the real entry name at the top of each created note.
+
 # 0.3.0
 Get detection of @JournalEntry working correctly again since the GFM module escapes the square brackets.